+++ so-if-db/src/main/java/com/pineone/icbms/so/interfaces/database/repository/VirtualObjectRepository.java
@@ -13,6 +13,6 @@ import javax.persistence.Table;
  */
 @Repository
 @Table(name = "virtual_object")
-public interface VirtualObjectRepository extends JpaRepository<VirtualObjectForDB, Integer> {
-    VirtualObjectForDB findById(int id);
+public interface VirtualObjectRepository extends JpaRepository<VirtualObjectForDB, Long> {
+    VirtualObjectForDB findById(long id);
 }
